+++ lib/provider/opencv/find-edges.function.ts
@@ -1,4 +1,4 @@
-import * as cv from "opencv4nodejs-prebuilt";
+import * as cv from "opencv4nodejs";
 
 export async function findEdges(image: cv.Mat): Promise<cv.Mat> {
   const gray = await image.cvtColorAsync(cv.COLOR_BGR2GRAY);