+++ CHANGELOG.md
@@ -48,4 +48,4 @@ All notable changes to this project will be documented in this file.
 
 ## 0.1.0-beta.1
 
-- Enabled pre-built OpenCV bindings via `opencv4nodejs-prebuilt`
+- Enabled pre-built OpenCV bindings via `opencv4nodejs`